DDF-4241 moved servlet filters to new bundle (#3912)
(cherry picked from commit 5d5ef52729d309ab69713e071ae15375eada7234)
