+++ platform/platform-paxweb-jettyconfig/pom.xml
@@ -93,6 +93,12 @@
             <version>${spring-osgi.version}</version>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>ddf.lib</groupId>
+            <artifactId>spock-shaded</artifactId>
+            <version>${project.version}</version>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
     <build>
         <plugins>
