+++ platform/platform-paxweb-jettyconfig/src/main/java/org/codice/ddf/pax/web/jetty/ClientInfoFilter.java
@@ -11,18 +11,19 @@
  * License is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package org.codice.ddf.platform.filter.clientinfo;
+package org.codice.ddf.pax.web.jetty;
 
 import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
+import javax.servlet.Filter;
+import javax.servlet.FilterChain;
+import javax.servlet.FilterConfig;
 import javax.servlet.ServletContext;
+import javax.servlet.ServletException;
 import javax.servlet.ServletRequest;
 import javax.servlet.ServletResponse;
 import org.apache.shiro.util.ThreadContext;
-import org.codice.ddf.platform.filter.AuthenticationException;
-import org.codice.ddf.platform.filter.FilterChain;
-import org.codice.ddf.platform.filter.SecurityFilter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -32,17 +33,17 @@ import org.slf4j.LoggerFactory;
  * for working with the client info map.
  *
  * <p>The information currently all comes from the servlet API; specifically a select few getters
- * within {@link javax.servlet.ServletRequest}. The format of the keys follows the format of java
- * beans. The keys are camel-cased names without the preceeding 'get' found in the method name.
+ * within {@link ServletRequest}. The format of the keys follows the format of java beans. The keys
+ * are camel-cased names without the preceeding 'get' found in the method name.
  *
- * <p>For example, the key associated with {@link javax.servlet.ServletRequest#getRemoteAddr()}
- * would be the string {@code remoteAddr}.
+ * <p>For example, the key associated with {@link ServletRequest#getRemoteAddr()} would be the
+ * string {@code remoteAddr}.
  *
  * <p>The only exception to this rule, {@link ClientInfoFilter#CLIENT_INFO_KEY}, which holds a value
  * string of {@code client-info}, is the key used to access the entire client information map. It
  * may contain different kinds of data that does not necessarily correlate to the servlet API.
  */
-public class ClientInfoFilter implements SecurityFilter {
+public class ClientInfoFilter implements Filter {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(ClientInfoFilter.class);
 
@@ -57,9 +58,7 @@ public class ClientInfoFilter implements SecurityFilter {
   public static final String SERVLET_CONTEXT_PATH = "contextPath";
 
   @Override
-  public void init() {
-    LOGGER.trace("Initializing Client Info Filter.");
-  }
+  public void init(FilterConfig filterConfig) throws ServletException {}
 
   @Override
   public void destroy() {}
@@ -67,7 +66,7 @@ public class ClientInfoFilter implements SecurityFilter {
   @Override
   public void doFilter(
       ServletRequest servletRequest, ServletResponse servletResponse, FilterChain filterChain)
-      throws IOException, AuthenticationException {
+      throws IOException, ServletException {
     ThreadContext.put(CLIENT_INFO_KEY, createClientInfoMap(servletRequest));
     try {
       filterChain.doFilter(servletRequest, servletResponse);
