+++ platform/platform-paxweb-jettyconfig/src/main/java/org/codice/ddf/pax/web/jetty/CsrfFilter.java
@@ -11,7 +11,7 @@
  * License is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package org.codice.ddf.security.filter.csrf;
+package org.codice.ddf.pax.web.jetty;
 
 import ddf.security.common.audit.SecurityLogger;
 import java.io.IOException;
@@ -20,6 +20,9 @@ import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.regex.Pattern;
+import javax.servlet.Filter;
+import javax.servlet.FilterChain;
+import javax.servlet.FilterConfig;
 import javax.servlet.ServletException;
 import javax.servlet.ServletRequest;
 import javax.servlet.ServletResponse;
@@ -30,16 +33,12 @@ import javax.ws.rs.core.MultivaluedHashMap;
 import javax.ws.rs.core.MultivaluedMap;
 import org.apache.commons.lang.StringUtils;
 import org.codice.ddf.configuration.SystemBaseUrl;
-import org.codice.ddf.platform.filter.AuthenticationChallengeException;
-import org.codice.ddf.platform.filter.AuthenticationException;
-import org.codice.ddf.platform.filter.FilterChain;
-import org.codice.ddf.platform.filter.SecurityFilter;
 import org.eclipse.jetty.http.HttpMethod;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /** Contains multiple checks to prevent cross-site requests for protected context paths. */
-public class CsrfFilter implements SecurityFilter {
+public class CsrfFilter implements Filter {
 
   public static final String CSRF_HEADER = "X-Requested-With";
   public static final String ORIGIN_HEADER = "Origin";
@@ -75,7 +74,7 @@ public class CsrfFilter implements SecurityFilter {
   }
 
   @Override
-  public void init() {
+  public void init(FilterConfig filterConfig) throws ServletException {
     LOGGER.debug("Starting CSRF filter.");
 
     // internal authority system properties
@@ -120,7 +119,7 @@ public class CsrfFilter implements SecurityFilter {
    */
   @Override
   public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain)
-      throws IOException, AuthenticationException {
+      throws IOException, ServletException {
     LOGGER.debug("Performing doFilter() on CsrfFilter");
     HttpServletRequest httpRequest = (HttpServletRequest) request;
     HttpServletResponse httpResponse = (HttpServletResponse) response;
@@ -131,15 +130,14 @@ public class CsrfFilter implements SecurityFilter {
     // Begin CSRF checks if request is accessing a Cross-Site protected context
     if (protectedContexts.stream().anyMatch(targetContextPath::startsWith)
         && doBrowserProtectionFilter(httpRequest, httpResponse, targetContextPath, requestMethod)) {
-      throw new AuthenticationChallengeException(
-          "Request is accessing a Cross-Site protected context");
+      return;
     }
 
     // Execute CSRF check if user is accessing /services
     if (targetContextPath.startsWith(SERVICE_CONTEXT)
         && doSystemProtectionFilter(
             httpRequest, httpResponse, targetContextPath, requestMethod, userAgentHeader)) {
-      throw new AuthenticationChallengeException("User is accessing /services");
+      return;
     }
 
     // All checks passed
