+++ platform/platform-paxweb-jettyconfig/src/main/resources/OSGI-INF/metatype/metatype.xml
