+++ platform/pom.xml
@@ -377,8 +377,6 @@
         <module>landing-page</module>
         <module>country-converter</module>
         <module>usng4j</module>
-        <module>platform-filter-clientinfo</module>
-        <module>platform-filter-response</module>
         <module>email</module>
         <module>osgi</module>
         <module>branding-registry-impl</module>
