+++ platform/security/filter/pom.xml
@@ -37,6 +37,5 @@
         <module>security-filter-web-sso</module>
         <module>security-filter-login</module>
         <module>security-filter-authorization</module>
-        <module>security-filter-csrf</module>
     </modules>
 </project>
