--- platform/security/filter/security-filter-csrf/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -1,56 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!-- /**
- * Copyright (c) Codice Foundation
- *
- * This is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser General Public License as published by the Free Software Foundation, either
- * version 3 of the License, or any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
- * See the GNU Lesser General Public License for more details. A copy of the GNU Lesser General Public License is distributed along with this program and can be found at
- * <http://www.gnu.org/licenses/lgpl.html>.
- *
- **/ -->
-<blueprint xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-           xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
-           xsi:schemaLocation="
-  http://www.osgi.org/xmlns/blueprint/v1.0.0 http://www.osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd">
-
-    <bean id="filter" class="org.codice.ddf.security.filter.csrf.CsrfFilter">
-        <property name="whiteListContexts">
-            <array value-type="java.lang.String">
-                <value>/services[/]?$=GET</value>
-                <value>/services/admin/config[/]?$=GET</value>
-                <value>/services/content[/]?$=GET</value>
-                <value>/services/catalog/query[/]?$=GET</value> <!-- Simple Search -->
-                <value>/services/catalog/sources.*=GET</value> <!-- REST Action providers -->
-                <value>/services/internal/catalog/download/cache.*=GET</value>
-                <value>/services/internal/session.*=GET</value>
-                <value>/services/internal/metrics.*=GET</value>
-                <value>/services/idp/login[/]?$=GET</value>
-                <value>/services/idp/login[/]?$=POST</value>
-                <value>/services/idp/login/metadata[/]?$=GET</value>
-                <value>/services/idp/login/sso[/]?$=GET</value>
-                <value>/services/idp/logout[/]?$=GET</value>
-                <value>/services/login[/]?$=POST</value>
-                <value>/services/logout/actions[/]?$=GET</value>
-                <value>/services/logout/local[/]?$=GET</value>
-                <value>/services/platform/config/ui[/]?$=GET</value>
-                <value>/services/saml/logout[/]?$=GET</value>
-                <value>/services/saml/logout/request[/]?$=GET</value>
-                <value>/services/saml/sso[/]?$=POST</value>
-                <value>/services/saml/sso/metadata[/]?$=GET</value>
-                <value>/services/SecurityTokenService[/]?$=GET</value>
-                <value>/services/SecurityTokenService[/]?$=POST</value>
-                <value>/services/store/config[/]?$=GET</value>
-            </array>
-        </property>
-    </bean>
-
-    <service ref="filter"
-             interface="org.codice.ddf.platform.filter.SecurityFilter" ranking="100">
-        <service-properties>
-            <entry key="osgi.http.whiteboard.filter.name" value="csrf-filter"/>
-        </service-properties>
-    </service>
-
-</blueprint>
